requirements.txt
PyQt5==5.15.9
cryptography==41.0.3
argon2-cffi==23.1.0 